refactor: put 'isLoggedIn' flag directy in store
